+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Removed 1.21.10 support
 - Added Hotspot Pointer with customizable priority.
   - Use at your own risk! Off by default
+- Added hotspot sharing
 
 ### Changes: